+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Release date `UNRELEASED`
 
+- (feat): implement `hsts` response header
+
 ## `1.0.0-28`
 
 Release date `2023-03-15`